debugging

+++ trace/dep_tracker.sh
@@ -2,7 +2,9 @@
 #strace v4.5.18 doesn't propagate the error code of the traced process. Always 0
 strace -f -s 1 -e trace=read,write,open -o $GENDEP_TARGET.trace $@
 if [ $? -eq 0 ]; then 
-	rm $GENDEP_TARGET.dep
+	rm -f $GENDEP_TARGET.dep
 	strace_log_filter.py $GENDEP_TARGET.trace $GENDEP_TARGET "$GENDEP_FMATCH"
-	rm $GENDEP_TARGET.trace
+	if [ "$GENDEP_DEBUG" != "1" ]; then
+		rm $GENDEP_TARGET.trace
+	fi
 fi
